move CreateOrUpdateWork() and related functions to controllers/ctrlutil
